refactor(sys): 重构系统管理模块
- 删除了冗余的模型定义文件 models.py
- 重新定义了部门、用户、角色等模型结构
- 优化了 CRUD 和服务层代码
- 统一了数据传输对象 (DTO) 的定义
- 移除了未使用的导入和注释

+++ src/common/base_api.py
@@ -1,16 +1,19 @@
 import asyncio
 
 from enum import Enum
-from typing import Any, Callable, Generic, Type
+from typing import Any, Callable, Generic, Sequence, Type
 
 from aiocache import RedisCache, cached
 from aiocache.serializers import PickleSerializer
 from fastapi import APIRouter, Body, Path, Query, Request
 from fastapi.params import Depends
+from sqlmodel import SQLModel
 from typing_extensions import Annotated
 
-from src.common.data_model.base_schema import BaseSchemaType, CreateSchemaType, ModelType, UpdateSchemaType, WithSchemaType
+from src.common.base_crud import CreateModelType, ModelType, UpdateModelType
+from src.common.base_service import BaseService
 from src.common.data_model.query_fields import QueryOptions
+from src.core.conf import settings
 from src.core.responses.response_schema import ResponseModel, response_base
 from src.database.cache.cache_conf import generate_cache_key, get_redis_settings
 from src.database.cache.cache_plugins import CacheLogPlugin
@@ -19,16 +22,17 @@ from src.database.db_session import CurrentSession, async_audit_session, async_s
 from src.database.redis_utils import RedisManager
 
 
-class BaseAPI(Generic[ModelType, CreateSchemaType, UpdateSchemaType]):
+class BaseAPI(Generic[ModelType, CreateModelType, UpdateModelType]):
     """基础 API 生成器"""
 
     def __init__(
         self,
         model: Type[ModelType],
-        create_schema: Type[CreateSchemaType] | None = None,
-        update_schema: Type[UpdateSchemaType] | None = None,
-        base_schema: Type[BaseSchemaType] | Any = Any,
-        with_schema: Type[WithSchemaType] | Any = Any,
+        service: BaseService[ModelType, CreateModelType, UpdateModelType],
+        create_schema: Type[CreateModelType] | None = None,
+        update_schema: Type[UpdateModelType] | None = None,
+        base_schema: Type[ModelType] | Any = Any,
+        with_schema: Type[ModelType] | Any = Any,
         prefix: str = "",
         tags: list[str | Enum] | None = None,
         dependencies: list[Depends] | None = None,
@@ -38,10 +42,11 @@ class BaseAPI(Generic[ModelType, CreateSchemaType, UpdateSchemaType]):
         gen_delete: bool = True,
         gen_bulk_delete: bool = False,
         gen_query: bool = True,
-        cache_ttl: int = 3600,
-        **kwargs: Any
+        cache_ttl: int = 3600 if settings.APP_ENV == "prod" else 60,
+        **kwargs: Any,
     ) -> None:
         self.model = model
+        self.service = service
         self.create_schema = create_schema
         self.update_schema = update_schema
         self.base_schema = base_schema
@@ -81,7 +86,6 @@ class BaseAPI(Generic[ModelType, CreateSchemaType, UpdateSchemaType]):
             self._register_query()
         # 基础查询接口总是生成
         self._register_get()
-        self._register_get_all()
 
     def _register_create(self) -> None:
         """注册创建接口"""
@@ -91,25 +95,25 @@ class BaseAPI(Generic[ModelType, CreateSchemaType, UpdateSchemaType]):
         )
         async def create(
             request: Request,
-            data: self.create_schema  # type: ignore
+            obj_in: Annotated[self.create_schema, Body(..., description="创建模型")]  # type: ignore
         ) -> ResponseModel[self.base_schema]:  # type: ignore
             async with async_audit_session(async_session(), request) as session:
-                result = await self.model.create(session=session, **data.model_dump(exclude_unset=True))
+                result = await self.service.create(session=session, obj_in=obj_in)
             return response_base.success(data=result)
 
     def _register_bulk_create(self) -> None:
-        """注册指创建接口"""
+        """注册批量创建接口"""
         @self.router.post(
             "/bulk_create",
             summary=f"批量创建{self.model.__name__}"
         )
         async def bulk_create(
             request: Request,
-            datas: Annotated[list[self.create_schema], Body(..., description="批量创建模型")]  # type: ignore
-        ) -> ResponseModel[list[self.base_schema]]:  # type: ignore
+            datas: Annotated[Sequence[self.create_schema], Body(..., description="批量创建模型")]  # type: ignore
+        ) -> ResponseModel[Sequence[self.base_schema]]:  # type: ignore
             async with async_audit_session(async_session(), request) as session:
-                result = await self.model.bulk_create(
-                    session=session, items=[data.model_dump() for data in datas]
+                result = await self.service.bulk_create(
+                    session=session, objects=datas
                 )
             return response_base.success(data=result)
 
@@ -121,13 +125,15 @@ class BaseAPI(Generic[ModelType, CreateSchemaType, UpdateSchemaType]):
         )
         async def update(
             request: Request,
-            data: Annotated[self.update_schema, Body(..., description="更新模型")]  # type: ignore
+            obj_in: Annotated[self.update_schema, Body(..., description="更新模型")]  # type: ignore
         ) -> ResponseModel[self.base_schema]:  # type: ignore
-            key = generate_cache_key(self.cache_prefix, f"id_{data.id}")
+            if not hasattr(obj_in, 'id'):
+                return response_base.fail(data="请求参数错误，ID不存在！")
+            key = generate_cache_key(self.cache_prefix, f"id_{getattr(obj_in, 'id')}")
             await redis_client.delete_prefix(key)
 
             async with async_audit_session(async_session(), request) as session:
-                result = await self.model.update(session=session, **data.model_dump(exclude_unset=True))
+                result = await self.service.update(session=session, obj_in=obj_in)
 
             # 延迟50ms后再次删除缓存，防止并发删除缓存时出现缓存不一致
             await redis_client.delete_prefix(key)
@@ -150,7 +156,7 @@ class BaseAPI(Generic[ModelType, CreateSchemaType, UpdateSchemaType]):
             await redis_client.delete_prefix(key)
 
             async with async_audit_session(async_session(), request) as session:
-                await self.model.delete(session=session, id=id)
+                await self.service.delete(session=session, id=id)
 
             # 延迟50ms后再次删除缓存，防止并发删除缓存时出现缓存不一致
             await asyncio.sleep(0.05)
@@ -174,13 +180,14 @@ class BaseAPI(Generic[ModelType, CreateSchemaType, UpdateSchemaType]):
                 await redis_client.delete_prefix(key)
 
             async with async_audit_session(async_session(), request) as session:
-                await self.model.bulk_delete(session=session, ids=ids)
-
+                unhandled_ids = await self.service.bulk_delete(session=session, ids=ids)
             # 延迟50ms后再次删除缓存
             await asyncio.sleep(0.05)
             for id in ids:
                 key = generate_cache_key(self.cache_prefix, f"id_{id}")
                 await redis_client.delete_prefix(key)
+            if unhandled_ids:
+                return response_base.success(data=f"未成功删除的ID: {unhandled_ids}")
             return response_base.success(data="批量删除数据成功")
 
     def _register_get(self) -> None:
@@ -207,10 +214,10 @@ class BaseAPI(Generic[ModelType, CreateSchemaType, UpdateSchemaType]):
             id: int,
             max_depth: Annotated[int, Query(le=3, description="关联数据的最大深度")] = 2
         ) -> ResponseModel[self.with_schema]:  # type: ignore
-            item = await self.model.get_by_id(session=session, id=id)
+            item = await self.service.get_by_id(session=session, id=id)
             if not item:
                 return response_base.fail(data=f"{self.model.__name__}不存在")
-            data = await item.to_dict(max_depth=max_depth)
+            data = await item.to_api_dict(max_depth=max_depth)
             return response_base.success(data=data)
 
     def _register_query(self) -> None:
@@ -223,22 +230,8 @@ class BaseAPI(Generic[ModelType, CreateSchemaType, UpdateSchemaType]):
             session: CurrentSession,
             options: QueryOptions
         ) -> ResponseModel:  # type: ignore
-            items, total = await self.model.query_with_count(session=session, options=options)
+            items, total = await self.service.crud.query_with_count(session=session, options=options)
             return response_base.success(data={"total": total, "items": items})
-
-    def _register_get_all(self) -> None:
-        """注册获取所有接口"""
-        @self.router.get(
-            "/get_all",
-            summary=f"获取所有{self.model.__name__}"
-        )
-        async def get_all(
-            session: CurrentSession,
-            include_deleted: Annotated[bool, Query(description="是否包含已删除数据")] = False
-        ) -> ResponseModel[list[self.base_schema]]:  # type: ignore
-            items = await self.model.get_all(session=session, include_deleted=include_deleted)
-            data = [await item.to_api_dict() for item in items]
-            return response_base.success(data=data)
 
     def include_router(self, router: APIRouter) -> None:
         """将路由包含到其他路由器中"""
